+++ windi.config.ts
@@ -60,7 +60,7 @@ export default defineConfig({
       "border-0 text-inherit read-only:cursor-default",
     "solid-select-list":
       "absolute min-w-full bg-[inherit] color-[inherit] shadow-lg " +
-      "whitespace-nowrap rounded-sm mt-1 p-2 z-1",
+      "whitespace-nowrap rounded-sm mt-1 p-2 z-1 overflow-y-auto max-h-50vh",
     "solid-select-option":
       "px-4 py-2 cursor-default select-none " +
       "hover:bg-gray-200 data-focused:bg-gray-100 " +